refactor: criação de models para interfaces dos serviços
