+++ archspot-angular/src/app/core/models/directory.model.ts
@@ -1,0 +1,8 @@
+export interface DirectoryDTO {
+  id: number;
+  name: string;
+  creationDate: string;
+  type: string;
+  projectId: number;
+  parentDirectoryId?: number;
+}
