+++ archspot-angular/src/app/core/models/user-project.model.ts
@@ -1,0 +1,13 @@
+export interface UserProjectResponse {
+  id: number;
+  userId: number;
+  userName: string;
+  projectId: number;
+  projectName: string;
+  role: string;
+}
+
+export interface UserProjectRequest {
+  userId: number;
+  role: string;
+}
