+++ archspot-angular/src/app/core/services/installment.service.ts
@@ -2,30 +2,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
-
-export type PaymentMethod = 'PIX' | 'CREDIT_CARD' | 'DEBIT_CARD' | 'BOLETO' | 'CHECK' | 'CASH';
-export type PaymentStatus = 'PENDING' | 'PAID' | 'OVERDUE' | 'CANCELED';
-
-export interface InstallmentRequest {
-  projectId: number;
-  amount: number;
-  estimatedPaymentDate: string;
-  paymentStatus?: PaymentStatus;
-  description?: string | null;
-  paymentMethod?: PaymentMethod | null;
-  realPaymentDate?: string | null;
-}
-
-export interface InstallmentResponse {
-  id: number;
-  estimatedPaymentDate: string;
-  realPaymentDate: string | null;
-  paymentMethod: PaymentMethod | null;
-  paymentStatus: PaymentStatus;
-  amount: number;
-  description?: string | null;
-  projectId: number;
-}
+import { InstallmentRequest, InstallmentResponse, PaymentMethod } from '../models/payment.model';
 
 @Injectable({ providedIn: 'root' })
 export class InstallmentService {
