+++ archspot-angular/src/app/payments/create-sequence-installments-modal/create-sequence-installments-modal.component.ts
@@ -1,5 +1,6 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { InstallmentResponse, InstallmentService, PaymentMethod } from '../../core/services/installment.service';
+import { InstallmentService } from '../../core/services/installment.service';
+import { InstallmentResponse, PaymentMethod } from '../../core/models/payment.model';
 
 @Component({
   selector: 'app-create-sequence-installments-modal',
