+++ archspot-angular/src/app/shared/directory/directory.component.css
@@ -51,3 +51,16 @@
   background: #fafafa;
   padding: 8px 12px;
 }
+
+.context-menu {
+  position: fixed;
+  z-index: 1000;
+}
+.context-menu li {
+  padding: 4px 8px;
+  cursor: pointer;
+}
+.context-menu li:hover {
+  background-color: #f0f0f0;
+}
+
