+++ archspot-angular/src/app/shared/directory/directory.component.html
@@ -1,31 +1,34 @@
-<div class="directory">
-  <div class="directory-header" (click)="toggleExpand()">
-    <!-- Ícone da pasta -->
-    <img src="assets/img/icons/folder.svg" width="20" height="20" class="me-2 folder-icon" [class.open]="expanded" />
-
-    <strong class="flex-grow-1">{{ directory.name }}</strong>
-
-    <!-- Ações -->
-    <div class="actions">
-      <img src="assets/img/icons/pencil.svg" class="icon" title="Renomear diretório"
-        (click)="renomearDiretorio(); $event.stopPropagation()" />
-
-      <img src="assets/img/icons/trash3.svg" class="icon" title="Excluir diretório"
-        (click)="deletarDiretorio(); $event.stopPropagation()" />
-
-      <img *ngIf="directory.type === 'DOCUMENTS'" src="assets/img/icons/upload.svg" class="icon"
-        title="Enviar documento" (click)="uploadFile(); $event.stopPropagation()" />
+<nav class="mb-1 d-flex justify-content-between align-items-center">
+  <div class="nav nav-tabs fs-5 d-flex align-items-center flex-wrap">
+    <div class="position-relative" *ngFor="let dir of directories">
+      <button class="nav-link py-2 px-3" [class.active]="dir.id === activeTabId" type="button"
+        (click)="setActiveTab(dir)" (contextmenu)="openContextMenu($event, dir)">
+        {{ dir.name }}
+      </button>
     </div>
+
+    <a *ngIf="!isCustomerInProject" class="btn align-middle ms-2 p-0" (click)="createDirectory()"
+      title="Novo diretório">
+      <img src="/assets/img/icons/plus-circle.svg" alt="+" width="20" height="20">
+    </a>
   </div>
 
-  <!-- Conteúdo (tabela de documentos) -->
-  <div class="directory-content" *ngIf="expanded">
-    <app-table *ngIf="documents.length > 0" [directoryId]="directory.id"
-      [title]="'Documentos em ' + directory.name"></app-table>
-
-    <!-- Se não houver documentos -->
-    <div *ngIf="documents.length === 0" class="text-muted small ps-4 py-2">
-      Nenhum documento neste diretório.
-    </div>
+  <div *ngIf="!isCustomerInProject && activeTabId">
+    <button class="btn btn-sm btn-light d-flex align-items-center me-1" (click)="uploadFile()">
+      <img src="/assets/img/icons/upload.svg" width="16" height="16" class="me-1">
+      Fazer Upload
+    </button>
   </div>
+</nav>
+
+<div class="tab-content border border-top-0 p-2 pt-0 mt-0">
+  <app-table *ngIf="activeTabId" [directoryId]="activeTabId" [projectId]="projectId"></app-table>
+</div>
+
+<div class="context-menu" *ngIf="contextMenu.visible"
+  [ngStyle]="{ top: contextMenu.y + 'px', left: contextMenu.x + 'px' }">
+  <ul class="list-unstyled m-0 p-2 bg-white border rounded shadow">
+    <li (click)="editDirectoryName(contextMenu.dir!,$event)">Renomear</li>
+    <li (click)="deleteDirectory(contextMenu.dir!)">Excluir</li>
+  </ul>
 </div>
